Practica2.2

+++ lib/rgb/RGB.cpp
@@ -8,7 +8,7 @@ RGB::RGB(){
     this->b = 0;
 };
 
-RGB::RGB(int r, int g, int b) {
+RGB::RGB(float r, float g, float b) {
     this->r = r;
     this->g = g;
     this->b = b;
